backend/requirements.txt
fastapi[all]==0.109.0
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
icalendar==5.0.11
pydantic-settings==2.1.0
python-dotenv==1.0.0
uvicorn[standard]==0.27.0
slowapi==0.1.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.26.0
